Feat: goal endpoints

+++ Backend/Controllers/FamilyController.cs
@@ -13,7 +13,7 @@ public class FamilyController : ControllerBase
 {
     [HttpPost("create-family")]
     [Authorize]
-    public ActionResult CreateFamily(string familyName)
+    public ActionResult<Family> CreateFamily(string familyName)
     {
         var userId = Convert.ToInt32(User.FindFirstValue(ClaimTypes.NameIdentifier));
         var family = FamilyDataAccess.CreateFamily(familyName);
@@ -30,8 +30,15 @@ public class FamilyController : ControllerBase
     [Authorize]
     public ActionResult DeleteFamily(int familyId) {
         var userId = Convert.ToInt32(User.FindFirstValue(ClaimTypes.NameIdentifier));
+        var user = UserDataAccess.GetUser(userId);
+        if (user == null) return NotFound("User with this ID, not found");
+
+        var family = FamilyDataAccess.GetFamily(familyId);
+        if (family == null) return NotFound("Family with this ID, not found");
         
+        if (!user.IsAdult) return Unauthorized("Only an adult can delete a family");
         if (!FamilyDataAccess.DeleteFamily(familyId)) return Problem("Could not find family with that ID");
+
         return Ok();
     }
 
@@ -73,7 +80,7 @@ public class FamilyController : ControllerBase
     
     [HttpGet("get-family-members")]
     [Authorize]
-    public IActionResult GetFamilyMembers()
+    public ActionResult<List<User>> GetFamilyMembers()
     {
         var userId = Convert.ToInt32(User.FindFirstValue(ClaimTypes.NameIdentifier));
         var user = UserDataAccess.GetUser(userId);
@@ -86,7 +93,7 @@ public class FamilyController : ControllerBase
 
     [HttpGet("get-family-name")]
     [Authorize]
-    public IActionResult GetFamilyName()
+    public ActionResult<string> GetFamilyName()
     {
         var userId = Convert.ToInt32(User.FindFirstValue(ClaimTypes.NameIdentifier));
         var user = UserDataAccess.GetUser(userId);
@@ -100,7 +107,7 @@ public class FamilyController : ControllerBase
     // Invite codes = family codes  
     [HttpPost("generate-invite")]
     [Authorize]
-    public IActionResult GenerateInvite()
+    public ActionResult<string> GenerateInvite()
     {
         var userId = Convert.ToInt32(User.FindFirstValue(ClaimTypes.NameIdentifier));
         var user = UserDataAccess.GetUser(userId);
